Fix crash with new cert pinning dialog

+++ ChatSecure/Classes/Utilities/UIAlertController+ChatSecure.swift
@@ -14,7 +14,7 @@ public extension UIAlertController {
     /** Returns a cert-pinning alert if needed */
     public static func certificateWarningAlert(error: Error, saveHandler: @escaping (_ action: UIAlertAction) -> Void) -> UIAlertController? {
         let nsError = error as NSError
-        guard let errorCode = OTRXMPPErrorCode(rawValue: UInt(nsError.code)),
+        guard let errorCode = OTRXMPPErrorCode(rawValue: nsError.code),
             errorCode == .sslError,
             let certData = nsError.userInfo[OTRXMPPSSLCertificateDataKey] as? Data,
             let hostname = nsError.userInfo[OTRXMPPSSLHostnameKey] as? String,
